Fully enabled input/output file copying and assignment creation, see notes
Decoupled the file writing and copying logic.
When the user selects a input or output file, they are saved in temp variables.
Directory creation and input/output copying only happen on clicking save if user has set up Assignment Name and Output File.
No longer disable buttons, pressing save without meeting requirements prompts the user to enter the name and output file, if they ignore the warning the form closes as canceled.
Added MatchPercentage field to this form with a default value of 90%